Crate xcm_procedural

Procedural macros used in XCM.

Macros

impl_conversion_functions_for_junctions_v2

impl_conversion_functions_for_junctions_v3

impl_conversion_functions_for_junctions_v4

impl_conversion_functions_for_location_v4

impl_conversion_functions_for_multilocation_v2

impl_conversion_functions_for_multilocation_v3

Derive Macros

Builder

This is called on the Instruction enum, not on the Xcm struct, and allows for the following syntax for building XCMs: let message = Xcm::builder() .withdraw_asset(assets) .buy_execution(fees, weight_limit) .deposit_asset(assets, beneficiary) .build();

XcmWeightInfoTrait